Cambiamos los salarios a integers
Si todos los salarios son enteros, no tiene caso estarlos
typecasteando a floats. Este cambio hace que se guarden como
enteros.

+++ lib/scraper_talento/candidate_processor.rb
@@ -72,7 +72,7 @@ module CandidateProcessor
   end
 
   def money(selector)
-    prop(selector).gsub(/[\s\$\,]/, '').to_f
+    prop(selector).gsub(/[\s\$\,]/, '').to_i
   end
 
   def name
